frontend/src/flask_api/requirements.txt
Flask>=2.0
flask-cors>=3.0
python-dotenv>=0.21
PyMySQL>=1.0
